Introduce wait modal
